// src/app/models/LotoFacil.ts
import Sorteo from "./Sorteo";

export default class Lotofacil extends Sorteo {
    
    bola1: string;
    bola2: string;
    bola3: string;
    bola4: string;
    bola5: string;
    bola6: string;
    bola7: string;
    bola8: string;
    bola9: string;
    bola10: string;
    bola11: string;
    bola12: string;
    bola13: string;
    bola14: string;
    bola15: string;    
}